+++ Hooks/InitializationPatch.cs
@@ -1,0 +1,19 @@
+using HarmonyLib;
+using Unity.Scenes;
+
+namespace CrimsonLog.Hooks;
+
+[HarmonyPatch]
+internal static class InitializationPatch
+{
+    [HarmonyPatch(typeof(SceneSystem), nameof(SceneSystem.ShutdownStreamingSupport))]
+    [HarmonyPostfix]
+    static void ShutdownStreamingSupportPostfix()
+    {
+        Core.InitializeAfterLoaded();
+        if (Core._hasInitialized)
+        {
+            Plugin.Harmony.Unpatch(typeof(SceneSystem).GetMethod("ShutdownStreamingSupport"), typeof(InitializationPatch).GetMethod("OneShot_AfterLoad_InitializationPatch"));
+        }
+    }
+}